requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
pypdf==4.3.1
sentence-transformers==3.0.1
faiss-cpu==1.13.2
streamlit==1.37.1
python-dotenv==1.0.1